+++ Erste Aufgaben in Java/ring.java
@@ -10,8 +10,8 @@ public class ring {
         
         System.out.println("Kreisring Recher:");
         System.out.println("Bekannt Variable: äußerem umd inneren Radius"); 
-        System.out.println("äußerer Radius: "); r1 = sc.nextInt();
-        System.out.println("inneren Radius: "); r2 = sc.nextInt();
+        System.out.println("äußerer Radius: "); r1 = sc.nextDouble();
+        System.out.println("inneren Radius: "); r2 = sc.nextDouble();
 
         A = Math.sqrt(Math.pow(r1, 2)-Math.pow(r2, 2));
